requirements.txt
Flask
openai
netmiko
python-dotenv